med-saas/backend/requirements.txt
fastapi>=0.116.0,<1
psycopg2-binary>=2.9.9,<3
python-multipart>=0.0.20,<1
rapidocr-onnxruntime>=1.2.3,<2
uvicorn>=0.35.0,<1
ultralytics>=8.3.0,<9
